requirements.txt

flask
metaapi-cloud-sdk
